% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cases.R
\name{filter_cases}
\alias{filter_cases}
\title{Filters cases from a MADS dataframe}
\usage{
filter_cases(data, ..., min.days.to.new.episode = 14)
}
\arguments{
\item{data}{a MADS dataframe}

\item{min.days.to.new.episode}{Minimum number of days between to episodes for it to be considered a new episode}

\item{positive}{a logical vector indicating positive samples}
}
\description{
Simplifies counting patients instead of samples
}
\keyword{MADS}
\keyword{cases}
\keyword{filter}

