% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MADS_like_data.R
\name{generate_MADS_like_data}
\alias{generate_MADS_like_data}
\title{Generating random MADS-like data}
\usage{
generate_MADS_like_data(model = "analyser")
}
\arguments{
\item{model}{either \code{analyser} or \code{species+res}}
}
\description{
Outputs a randomly generated dataframe of MADS-like data into ./data/MADS-like.csv.
Probably only useful for testing.
}

