% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_res.R
\name{generate_res}
\alias{generate_res}
\title{Generating resistances by bacterium}
\usage{
generate_res(bakterier)
}
\arguments{
\item{bakterier}{a vector of bacterial names.}
}
\description{
Takes a vector of bacterial names, outputs a dataframe of simulated resistances
Intended for use with generate_MADS_like_data, specifically in the species+res model
}

