% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mads.R
\name{read_mads}
\alias{read_mads}
\title{Read MADS output into a dataframe}
\usage{
read_mads(file, model)
}
\arguments{
\item{file}{the name of the file to be read from.}

\item{model}{Which model was the MADS file made with?}
}
\description{
Read MADS output into a dataframe
}
\keyword{MADS}
\keyword{read}

