% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slplot.R
\name{slplot}
\alias{slplot}
\title{Provide a "since last new case" plot}
\usage{
slplot(sldata)
}
\arguments{
\item{sldata}{a "since last new case" dataframe}
}
\description{
Counts on a patient, not a sample, basis.
}
\keyword{MADS}
\keyword{last}
\keyword{plot}
\keyword{since}

