% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_by_department.R
\name{tally_by_department}
\alias{tally_by_department}
\title{Tally positive and negative cases by department}
\usage{
tally_by_department(data, basis, ...)
}
\arguments{
\item{data}{a MADS dataframe}

\item{basis}{Two bases: "Patient" which counts by patient, such that any patient with one positive test will be counted as positive. "Sample": Samples are counted individually}

\item{positive}{a logical vector indicating positive samples}
}
\description{
Patient count: counts such that any patient which one positive test will be counted as positive
}
\details{
Sample count: counts individual samples
}
\keyword{MADS}
\keyword{count}
\keyword{tally}

