% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_map.R
\name{tally_map}
\alias{tally_map}
\title{Create a heatmap, intented for use with tally_by_department}
\usage{
tally_map(data)
}
\arguments{
\item{data}{a dataframe to be plotted, expects values hosp_afd and either month or week}

\item{...}{used for providing a window, either month or week, defaults to month}
}
\description{
A convenience function that provides access to a specific ggplot
}
\keyword{MADS}
\keyword{count}
\keyword{heatmap}
\keyword{tally}

