% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincle_last.R
\name{since_last}
\alias{since_last}
\title{Provide a "since last new case" dataframe}
\usage{
since_last(data)
}
\arguments{
\item{data}{a MADS dataframe}

\item{positive}{a logical vector indicating positive samples}

\item{min.days.to.new.episode}{Minimum number of days between to episodes for it to be considered a new episode}
}
\description{
Counts on a patient, not a sample, basis.
}
\keyword{MADS}
\keyword{last}
\keyword{since}

